pyproject.toml
[project]
name = "llm-pygame"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "pygame>=2.6.1",
    "openai>=1.0.0",
    "python-dotenv>=1.0.0",
]
